refactor(recorder): move browser-watchdog to failure/
Slice #10 introduces a new failure/ home for the three failure-mode
watchdogs (device, crash, browser). Browser-watchdog moves from
server/ to failure/ unchanged in behavior — same class, same tests.

Refs #21
Refs #31
