+++ tools/recorder/src/failure/browser-watchdog.js
